requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.22
python-dotenv>=1.0.0
pydantic>=2.0.0
tldextract>=5.0.0
sentence-transformers>=2.2.0
